Workaround across SqlValue and Enum handling.
